Remove parser-less support from ob_parse_arguments_list()
CMake doesn't properly support a list of lists, and the implementation
of appending entry words to the entry arguments list of this function
was an attempt to work around that; however, due to how CMake processes
function arguments, the resultant entry list was being passed as one
argument, even though it wasn't quoted in the parser function call.

It seems there is no viable way to handle this other than making it a
separate function, but event the user would have to manually replace
the whitespace of each list entry with semi-colons when the were to
use each entry, which is quite clunky.

For now it's better to just drop the concept.