Revert separate debug executable install
The approach didn't fully work since the dependency install code
wasn't also marked as optional, and actually can't be marked as
optional. The custom code could be wraped in a check to see if the
installed executable target exists, but then there is the issue
of the qt deploy script which cannot be wrapped and also can't
be marked as optional.

The best way to handle the situation this was designed for,
building only debug libs but not executables and then installing
the debug configuration, is to also specify the component of the
libs when invoking install for the debug configuration.