Remove check for system zlib from FetchQuaZip, handle shared builds
FindZLIB has an issue in that it doesn't always respect the requested
library type, meaning one can't be sure if the found zlib is shared
or static. We want to guarantee it's one or the other so that means
we must always build from source until CMake improves the situation.

See the included note.