Add typical rpath (runpath) defaults to top-level projects
Adds $ORIGIN and $ORIGIN/../lib to the rpath of all targets by default
for top-level projects to better ensure their dependencies can be
found at runtime. We set the build and install tree values so that in
the case where a target is simply copied into the install tree and not
properly installed (which unfortunately is the workaround used via the
Executable module, due to CMake not having a convenient way to install
ONLY the runtime dependencies of in-tree targets, instead relying on
their own install rules having a way to do so, which makes installing
them for FetchContent based targets unreliable since most projects do
not have this), they will still have appropriate rpath values in the
install tree.